+++ src/globalCallables.ts
@@ -10,12 +10,10 @@ import { InterfaceType } from './types/InterfaceType';
 import { ObjectType } from './types/ObjectType';
 import { StringType } from './types/StringType';
 import { VoidType } from './types/VoidType';
-import { Parser } from './parser';
 import util from './util';
 
 export let globalFile = new BrsFile('global', 'global', null);
-globalFile.parser = new Parser();
-globalFile.parser.parse([]);
+globalFile.parse('');
 
 let mathFunctions = [{
     name: 'Abs',